+++ test/apple-pay/unit/apple-pay.js
@@ -22,8 +22,26 @@ describe('ApplePay', function () {
     });
   });
 
-  describe('decoratePaymentRequest', function () {
+  it('exposes a readonly merchantIdentifier property', function () {
+    var err;
+
+    expect(this.applePay.merchantIdentifier).to.equal('com.example.test-merchant-identifier');
+
+    try {
+      this.applePay.merchantIdentifier = 'something wrong';
+    } catch (e) {
+      err = e;
+    }
+
+    expect(err).to.be.an.instanceof(Error);
+    expect(err.message).to.equal('Attempted to assign to readonly property.');
+    expect(this.applePay.merchantIdentifier).to.equal('com.example.test-merchant-identifier');
+  });
+
+  describe('createPaymentRequest', function () {
     beforeEach(function () {
+      var self = this;
+
       this.paymentRequest = {
         countryCode: 'defined',
         currencyCode: 'defined',
@@ -31,50 +49,67 @@ describe('ApplePay', function () {
         supportedNetworks: ['defined']
       };
 
+      this.gatewayConfiguration = {
+        applePay: {
+          countryCode: 'decorated',
+          currencyCode: 'decorated',
+          merchantCapabilities: ['decorated'],
+          supportedNetworks: ['visa', 'amex', 'mastercard']
+        }
+      };
+
       this.client.getConfiguration = function () {
-        return {
-          gatewayConfiguration: {
-            applePay: {
-              countryCode: 'decorated',
-              currencyCode: 'decorated',
-              merchantCapabilities: ['decorated'],
-              supportedNetworks: ['visa', 'amex', 'mastercard']
-            }
-          }
-        };
+        return {gatewayConfiguration: self.gatewayConfiguration};
       };
     });
 
+    it("doesn't mutate the argument", function () {
+      var arg = {foo: 'boo'};
+
+      this.applePay.createPaymentRequest(arg);
+      expect(arg).to.deep.equal({foo: 'boo'});
+    });
+
+    it('leaves properties intact', function () {
+      expect(this.applePay.createPaymentRequest({foo: 'boo'}).foo).to.equal('boo');
+    });
+
     it('applies countryCode when undefined', function () {
-      expect(this.applePay.decoratePaymentRequest({}).countryCode).to.equal('decorated');
+      expect(this.applePay.createPaymentRequest({}).countryCode).to.equal('decorated');
     });
 
     it('applies currencyCode when undefined', function () {
-      expect(this.applePay.decoratePaymentRequest({}).currencyCode).to.equal('decorated');
+      expect(this.applePay.createPaymentRequest({}).currencyCode).to.equal('decorated');
     });
 
-    it('applies merchantCapabilities when undefined', function () {
-      expect(this.applePay.decoratePaymentRequest({}).merchantCapabilities).to.eql(['decorated']);
+    it('applies gateway merchantCapabilities when not defined on argument but defined on gateway', function () {
+      expect(this.applePay.createPaymentRequest({}).merchantCapabilities).to.deep.equal(['decorated']);
+    });
+
+    it('applies default merchantCapabilities when not defined on argument and gateway', function () {
+      delete this.gatewayConfiguration.applePay.merchantCapabilities;
+
+      expect(this.applePay.createPaymentRequest({}).merchantCapabilities).to.deep.equal(['supports3DS']);
     });
 
     it('applies supportedNetworks when undefined', function () {
-      expect(this.applePay.decoratePaymentRequest({}).supportedNetworks).to.eql(['visa', 'amex', 'masterCard']);
+      expect(this.applePay.createPaymentRequest({}).supportedNetworks).to.deep.equal(['visa', 'amex', 'masterCard']);
     });
 
     it('does not apply countryCode when defined', function () {
-      expect(this.applePay.decoratePaymentRequest(this.paymentRequest).countryCode).to.equal('defined');
+      expect(this.applePay.createPaymentRequest(this.paymentRequest).countryCode).to.equal('defined');
     });
 
     it('does not apply currencyCode when defined', function () {
-      expect(this.applePay.decoratePaymentRequest(this.paymentRequest).currencyCode).to.equal('defined');
+      expect(this.applePay.createPaymentRequest(this.paymentRequest).currencyCode).to.equal('defined');
     });
 
     it('does not apply merchantCapabilities when defined', function () {
-      expect(this.applePay.decoratePaymentRequest(this.paymentRequest).merchantCapabilities).to.eql(['defined']);
+      expect(this.applePay.createPaymentRequest(this.paymentRequest).merchantCapabilities).to.deep.equal(['defined']);
     });
 
     it('does not apply supportedNetworks when defined', function () {
-      expect(this.applePay.decoratePaymentRequest(this.paymentRequest).supportedNetworks).to.eql(['defined']);
+      expect(this.applePay.createPaymentRequest(this.paymentRequest).supportedNetworks).to.deep.equal(['defined']);
     });
   });
 
@@ -88,7 +123,7 @@ describe('ApplePay', function () {
         err = e;
       }
 
-      expect(err).to.be.an.instanceOf(BraintreeError);
+      expect(err).to.be.an.instanceof(BraintreeError);
       expect(err.code).to.equal('CALLBACK_REQUIRED');
       expect(err.type).to.equal('MERCHANT');
       expect(err.message).to.equal('performValidation requires a callback.');
@@ -97,7 +132,7 @@ describe('ApplePay', function () {
     it('calls callback with error when event is undefined', function (done) {
       this.applePay.performValidation(false, function (err, data) {
         expect(data).not.to.exist;
-        expect(err).to.be.an.instanceOf(BraintreeError);
+        expect(err).to.be.an.instanceof(BraintreeError);
         expect(err.code).to.equal('APPLE_PAY_VALIDATION_URL_REQUIRED');
         expect(err.type).to.equal('MERCHANT');
         expect(err.message).to.equal('performValidation must be called with a validationURL.');
@@ -109,7 +144,7 @@ describe('ApplePay', function () {
     it('calls callback with error when event.validationURL is undefined', function (done) {
       this.applePay.performValidation({some: 'property'}, function (err, data) {
         expect(data).not.to.exist;
-        expect(err).to.be.an.instanceOf(BraintreeError);
+        expect(err).to.be.an.instanceof(BraintreeError);
         expect(err.code).to.equal('APPLE_PAY_VALIDATION_URL_REQUIRED');
         expect(err.type).to.equal('MERCHANT');
         expect(err.message).to.equal('performValidation must be called with a validationURL.');
@@ -133,8 +168,17 @@ describe('ApplePay', function () {
       }));
     });
 
-    it('makes a request to the gateway and calls the callback if it fails', function (done) {
-      var requestError = new Error('something went wrong');
+    it('makes a request to the gateway and calls the callback if it fails due to an invalid request', function (done) {
+      var merchantValidationError = new Error('Validation failed');
+      var requestError = new BraintreeError({
+        type: 'NETWORK',
+        code: 'CLIENT_REQUEST_ERROR',
+        message: 'This is a fake client error',
+        details: {
+          originalError: merchantValidationError
+        }
+      });
+
       var validationOptions = {
         validationURL: 'https://apple-pay-gateway-cert.apple.com/paymentservices/startSession',
         displayName: 'Awesome Merchant'
@@ -148,17 +192,53 @@ describe('ApplePay', function () {
         expect(options.data.applePayWebSession.domainName).to.equal('localhost');
         expect(options.data.applePayWebSession.displayName).to.equal('Awesome Merchant');
         expect(options.data.applePayWebSession.validationUrl).to.equal('https://apple-pay-gateway-cert.apple.com/paymentservices/startSession');
-        callback(new Error(requestError));
+        callback(requestError);
       };
 
       this.applePay.performValidation(validationOptions, function (err, response) {
         expect(response).not.to.exist;
 
-        expect(err).to.be.an.instanceOf(BraintreeError);
-        expect(err.code).to.equal('APPLE_PAY_MERCHANT_VALIDATION');
+        expect(err).to.be.an.instanceof(BraintreeError);
+        expect(err.code).to.equal('APPLE_PAY_MERCHANT_VALIDATION_FAILED');
+        expect(err.type).to.equal('MERCHANT');
+        expect(err.message).to.equal('Make sure you have registered your domain name in the Braintree Control Panel.');
+        expect(err.details.originalError.message).to.equal('Validation failed');
+
+        done();
+      });
+    });
+
+    it('makes a request to the gateway and calls the callback if it fails due to a network error', function (done) {
+      var requestError = new BraintreeError({
+        type: 'NETWORK',
+        code: 'CLIENT_REQUEST_TIMEOUT',
+        message: 'This is a fake client error'
+      });
+
+      var validationOptions = {
+        validationURL: 'https://apple-pay-gateway-cert.apple.com/paymentservices/startSession',
+        displayName: 'Awesome Merchant'
+      };
+
+      this.client.request = function (options, callback) {
+        expect(options.method).to.equal('post');
+        expect(options.endpoint).to.equal('apple_pay_web/sessions');
+        expect(options.data._meta.source).to.equal('apple-pay');
+        expect(options.data.applePayWebSession.merchantIdentifier).to.equal('com.example.test-merchant-identifier');
+        expect(options.data.applePayWebSession.domainName).to.equal('localhost');
+        expect(options.data.applePayWebSession.displayName).to.equal('Awesome Merchant');
+        expect(options.data.applePayWebSession.validationUrl).to.equal('https://apple-pay-gateway-cert.apple.com/paymentservices/startSession');
+        callback(requestError);
+      };
+
+      this.applePay.performValidation(validationOptions, function (err, response) {
+        expect(response).not.to.exist;
+
+        expect(err).to.be.an.instanceof(BraintreeError);
+        expect(err.code).to.equal('APPLE_PAY_MERCHANT_VALIDATION_NETWORK');
         expect(err.type).to.equal('NETWORK');
         expect(err.message).to.equal('A network error occurred when validating the Apple Pay merchant.');
-        expect(err.details).to.deep.equal({originalError: requestError});
+        expect(err.details.originalError.message).to.equal('This is a fake client error');
 
         done();
       });
@@ -203,7 +283,7 @@ describe('ApplePay', function () {
         err = e;
       }
 
-      expect(err).to.be.an.instanceOf(BraintreeError);
+      expect(err).to.be.an.instanceof(BraintreeError);
       expect(err.code).to.equal('CALLBACK_REQUIRED');
       expect(err.type).to.equal('MERCHANT');
       expect(err.message).to.equal('tokenize requires a callback.');
@@ -213,7 +293,7 @@ describe('ApplePay', function () {
       this.applePay.tokenize(false, function (err, paymentData) {
         expect(paymentData).not.to.exist;
 
-        expect(err).to.be.an.instanceOf(BraintreeError);
+        expect(err).to.be.an.instanceof(BraintreeError);
         expect(err.code).to.equal('APPLE_PAY_PAYMENT_TOKEN_REQUIRED');
         expect(err.type).to.equal('MERCHANT');
         expect(err.message).to.equal('tokenize must be called with a payment token.');
@@ -242,7 +322,7 @@ describe('ApplePay', function () {
       this.applePay.tokenize({token: token}, function (err, response) {
         expect(response).not.to.exist;
 
-        expect(err).to.be.an.instanceOf(BraintreeError);
+        expect(err).to.be.an.instanceof(BraintreeError);
         expect(err.code).to.equal('APPLE_PAY_TOKENIZATION');
         expect(err.type).to.equal('NETWORK');
         expect(err.message).to.equal('A network error occurred when processing the Apple Pay payment.');